Lift out all results handling code
