--- docs/source/api_reference/psimulate/utilities.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.psimulate.utilities
